Foi criado o modulo notes para organizar as features de notas do usuario
