+++ front/templates/dashboard/starter.html
@@ -48,7 +48,7 @@
               <span class="info-box-icon bg-danger"><i class="far fa-star"></i></span>
 
               <div class="info-box-content">
-                <span class="info-box-text"><a href="{{ url_for('listar_notas') }}">Notas</a></span>
+                <span class="info-box-text"><a href="{{ url_for('notes.listar_notas') }}">Notas</a></span>
                 <span class="info-box-number">{{ dadosDashboardDTO.quantidade_notas }}</span>
               </div>
               <!-- /.info-box-content -->
