+++ front/templates/notas/nova_nota.html
@@ -14,7 +14,7 @@
                     <div class="card-header">
                         <h3 class="card-title">Nota</h3>
                     </div>
-                    <form action="{{ url_for('nova_nota') }}" method="post">
+                    <form action="{{ url_for('notes.nova_nota') }}" method="post">
                         <div class="card-body">
                             <div class="form-group">
                                 <label for="descricaonota">Descrição</label>
